+++ detoxify/bias_metrics.py
@@ -1,6 +1,9 @@
 import pandas as pd
 import numpy as np
-from sklearn.metrics import roc_auc_score
+try:
+    from sklearn.metrics import roc_auc_score
+except ImportError:
+    print("Package `sklearn` is missing.")
 
 IDENTITY_COLUMNS = [
     "male",
